[Keyboard] Update MJ61 architecture to support multi version (#11787)
