+++ keyboards/melgeek/mj61/rev2/config.h
@@ -28,12 +28,12 @@
  */
 
 #define MATRIX_ROW_PINS { B12, B11, B10, B1, A3 }
-#define MATRIX_COL_PINS { B15, A8, A10, A15, B3, B4, B5, B8, B9, C13, C14, C15, A0, A1 }
+#define MATRIX_COL_PINS { B15, A8, B13, A15, B3, B4, B5, B8, B9, C13, C14, C15, A0, A1 }
 #define UNUSED_PINS
 
 /* COL2ROW, ROW2COL*/
 #define DIODE_DIRECTION COL2ROW
 
-#define DRIVER_1_LED_TOTAL 63
+#define DRIVER_1_LED_TOTAL 71
 #define DRIVER_LED_TOTAL DRIVER_1_LED_TOTAL
 #define DRIVER_INDICATOR_LED_TOTAL 0
