+++ keyboards/melgeek/mj61/rev2/rules.mk
@@ -1,6 +1,5 @@
 # MCU name
 MCU = STM32F303
-BOARD = QMK_PROTON_C
 
 # Build Options
 #   change yes to no to disable
